Commit CJL's version 'F' drivers.
This adds large blocks of comments at the front to describe the
revision history, as well as comments throughout to help explain
things.  Various tweaks for compatibility with various assembler's
quirks.  The device group names are tweaked so the system and
non-system handlers look related in BUILD, etc.

Control-C detection is tweaked to be compatible with DECmates.